% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_IEDB.R
\name{assemble_curlopts}
\alias{assemble_curlopts}
\title{Title}
\usage{
assemble_curlopts(
  protein_sequence,
  hla_alleles,
  prediction_method,
  peptide_length
)
}
\arguments{
\item{protein_sequence}{}

\item{hla_alleles}{}

\item{prediction_method}{}

\item{peptide_length}{}
}
\value{

}
\description{
Title
}
