% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{predict_mhc_peptides}
\alias{predict_mhc_peptides}
\title{Predict MHC binding peptides using IEDB analysis tools}
\usage{
predict_mhc_peptides(
  protein_input,
  hla_alleles,
  prediction_method = "consensus",
  peptide_length,
  mhc_class = "MHC-I"
)
}
\arguments{
\item{hla_alleles}{List of HLA alleles.}

\item{prediction_method}{Name of the prediction method to be used.}

\item{peptide_length}{List of peptide lengths (one HLA allele - one peptide length)}

\item{mhc_class}{MHC class. Currently only "MHC-I" is supported.}

\item{protein_sequence}{Protein sequence used for binding prediciton.}
}
\value{
dataframe
}
\description{
Predict MHC binding peptides using IEDB analysis tools
}
\examples{



}
